requirements.txt
asgiref
autopep8==1.5.7
certifi==2021.5.30
coverage==5.5
Django
django-cors-headers==3.8.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
psutil==5.8.0
psycopg2
pycodestyle==2.7.0
PyJWT==2.1.0
pytz
sqlparse
toml==0.10.2
typing-extensions
wincertstore==0.2
